Migraciones en Larevel, Modelos, Controladores y APIs basicas

+++ backend/app/Http/Controllers/CategoriaController.php
@@ -4,35 +4,87 @@ namespace App\Http\Controllers;
 
 use App\Models\Categoria;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Validator;
 
 class CategoriaController extends Controller
 {
     public function index()
     {
-        return Categoria::all();
+        $categorias = Categoria::all();
+
+        if ($categorias->isEmpty()) {
+            return response()->json(['message' => 'No se encontraron categorías'], 404);
+        }
+
+        return response()->json($categorias);
     }
 
+    // Método para crear una nueva categoría
     public function store(Request $request)
     {
-        return Categoria::create($request->all());
+        $validator = Validator::make($request->all(), [
+            'nombre_categoria' => 'required|string|max:25',
+            
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 422);
+        }
+
+        $categoria = Categoria::create([
+            'nombre_categoria' => $request->input('nombre_categoria'),
+            
+        ]);
+
+        return response()->json($categoria);
     }
 
     public function show($id)
     {
-        return Categoria::findOrFail($id);
+        $categoria = Categoria::find($id);
+
+        if (!$categoria) {
+            return response()->json(['message' => 'Categoría no encontrada'], 404);
+        }
+        return response()->json($categoria);
     }
 
     public function update(Request $request, $id)
     {
-        $categoria = Categoria::findOrFail($id);
-        $categoria->update($request->all());
-        return $categoria;
+        $categoria = Categoria::find($id);
+
+        if (!$categoria) {
+            return response()->json(['message' => 'Categoría no encontrada'], 404);
+        }
+
+        $validator = Validator::make($request->all(), [
+            'nombre_categoria' => 'required|string|max:25',     
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 422);
+        }
+
+        if ($request->has('nombre_categoria')) {
+            $categoria->nombre_categoria = $request->input('nombre_categoria');
+        }
+
+        $categoria->save();
+
+        return response()->json($categoria);
     }
 
     public function destroy($id)
     {
-        $categoria = Categoria::findOrFail($id);
+        $categoria = Categoria::find($id);
+
+        if (!$categoria) {
+            return response()->json(['message' => 'Categoría no encontrada'], 404);
+        }
+
         $categoria->delete();
-        return response(null, 204);
+
+        return response()->json(['message' => 'Categoría eliminada correctamente']);
     }
 }
